+++ ArgumentValidationFunctions/mustBeDoulbyPeriodicFPlane.m
@@ -1,3 +1,0 @@
-function mustBeDoulbyPeriodicFPlane(a)
-assert( isa(a,'WVGeometryDoublyPeriodic') && isa(a,'WVRotatingFPlane'),'mustBeDoulbyPeriodicFPlane::invalidClass','This geostrophic component is only valid for double periodic geometry on a rotating f-plane.');
-end