+++ Forcing/WVFixedAmplitudeForcing.m
@@ -134,7 +134,7 @@ classdef WVFixedAmplitudeForcing < WVForcing
             % set the amplitude to fix for the wave part of the flow
             %
             % This function will automatically remove modes set in the
-            % damping region of the WVAdapativeDamping forcing, if present.
+            % damping region of the WVAdaptiveDamping forcing, if present.
             %
             % - Topic: Setting the forcing
             % - Declaration:  setWaveForcingCoefficients(Apbar,Ambar,options)
@@ -175,7 +175,7 @@ classdef WVFixedAmplitudeForcing < WVForcing
             % set amplitude to fix for the geostrophic part of the flow
             %
             % This function will automatically remove modes set in the
-            % damping region of the WVAdapativeDamping forcing, if present.
+            % damping region of the WVAdaptiveDamping forcing, if present.
             %
             % - Topic: Setting the forcing
             % - Declaration: setGeostrophicForcingCoefficients(A0bar,options)