+++ Integrators/WVModelAdaptiveTimeStepCellMethods.m
@@ -1,4 +1,4 @@
-classdef WVModelAdapativeTimeStepCellMethods < handle
+classdef WVModelAdaptiveTimeStepCellMethods < handle
     %UNTITLED3 Summary of this class goes here
     %   Detailed explanation goes here